feat(api): 新增 CancellablePlugin 类型来定义插件的可取消行为；将部分类型定义从单一文件内拆除出去
