+++ simbot-api/src/commonMain/kotlin/love/forte/simbot/plugin/PluginFactory.kt
@@ -1,0 +1,46 @@
+/*
+ *     Copyright (c) 2026. ForteScarlet.
+ *
+ *     Project    https://github.com/simple-robot/simpler-robot
+ *     Email      ForteScarlet@163.com
+ *
+ *     This file is part of the Simple Robot Library (Alias: simple-robot, simbot, etc.).
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU Lesser General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     Lesser GNU General Public License for more details.
+ *
+ *     You should have received a copy of the Lesser GNU General Public License
+ *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ */
+
+package love.forte.simbot.plugin
+
+import love.forte.simbot.common.function.MergeableFactory
+
+/**
+ * [Plugin] 的工厂函数，用于配置并预构建 [Plugin] 实例。
+ *
+ * @see Plugin
+ * @param P 目标组件类型
+ * @param CONF 配置类型。配置类型应是一个可变类，以便于在 DSL 中进行动态配置。
+ */
+public interface PluginFactory<P : Plugin, CONF : Any> :
+    MergeableFactory<PluginFactory.Key, P, CONF, PluginConfigureContext> {
+    /**
+     * 用于 [PluginFactory] 在内部整合时的标识类型。
+     *
+     * 更多说明参阅 [MergeableFactory.Key]。
+     *
+     * @see PluginFactory.key
+     * @see MergeableFactory.key
+     */
+    public interface Key : MergeableFactory.Key
+}
